requirements.txt

fastapi==0.70.1
PyGithub==1.55
PyYAML==6.0
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
uvicorn==0.16.0
hvac==0.11.2